Refactor held-out benchmark ownership
